requirements.txt
flake8==3.8.2
Pillow
pyyaml
sentencepiece
tensorboard
tensorboardX
typeguard
huey
redis